pyproject.toml
[tool.poetry]
name = "hack-the-back"
version = "0.1.0"
description = ""
authors = ["Alvin Tang <alvin.tang@mail.utoronto.ca>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
django-ordered-model = "^3.4.3"
social-auth-app-django = "^4.0.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^5.1.0"}
djangorestframework = "^3.12.4"
PyJWT = "^1.7.1"
django-graphql-social-auth = {extras = ["jwt"], version = "^0.1.4"}
djangorestframework-stubs = "^1.4.0"
django-stubs = "^1.8.0"
django-graphql-jwt = "^0.3.1"
drf-spectacular = "^0.16.0"
djangorestframework-camel-case = "^1.2.0"
graphene-django = "^2.15.0"

[tool.poetry.dev-dependencies]
pylint = "^2.8.2"
black = "^21.5b1"
isort = "^5.8.0"
pytest = "^6.2.4"
pytest-django = "^4.3.0"
pylint-django = "^2.4.4"
rope = "^0.19.0"
freezegun = "^1.1.0"
Faker = "^8.1.4"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
